+++ Net_Core/Web.Support/Utilities/ClientChoicesUtility.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Data;
 using System.Text.Json;
+using System.Threading.Tasks;
 using GrowthWare.BusinessLogic;
 using GrowthWare.Framework;
 using GrowthWare.Framework.Models;
@@ -163,11 +164,11 @@ public static class ClientChoicesUtility
     /// <param name="clientChoicesState">MClientChoicesState</param>
     /// <param name="updateContext">bool</param>
     /// <remarks>Calls getClientChoicesState to ensure the Session/Cache matches for the given account.</remarks>
-    public static void Save(MClientChoicesState clientChoicesState)
+    public static async Task Save(MClientChoicesState clientChoicesState)
     {
         if (clientChoicesState == null) throw new ArgumentNullException(nameof(clientChoicesState), "clientChoicesState cannot be a null reference (Nothing in Visual Basic)! (Nothing in VB)!");
         BClientChoices mBusinessLayer = new BClientChoices(SecurityEntityUtility.DefaultProfile(), ConfigSettings.CentralManagement);
-        mBusinessLayer.Save(clientChoicesState);
+        await mBusinessLayer.Save(clientChoicesState);
         getClientChoicesState(clientChoicesState.Account);
     }