+++ Release_Notes.md
@@ -23,6 +23,9 @@
     - DFunctions
         Excluding:
                 - DataSet GetFunctions()
+    - DClientChoices
+        Excluding:
+                - DataRow GetChoices(string account)
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be, causing a logout to occure